refactor: split e2e tests
