--- test/e2e/diff_test.go
@@ -1,1034 +1,0 @@
-/*
-Copyright 2025 The Crossplane Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2e
-
-import (
-	"bufio"
-	"bytes"
-	"context"
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"regexp"
-	"slices"
-	"strings"
-	"testing"
-	"time"
-	"unicode"
-
-	corev1 "k8s.io/api/core/v1"
-	k8sapiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/e2e-framework/pkg/envconf"
-	"sigs.k8s.io/e2e-framework/pkg/features"
-
-	xpv1 "github.com/crossplane/crossplane-runtime/v2/apis/common/v1"
-	"github.com/crossplane/crossplane-runtime/v2/pkg/resource/unstructured/composed"
-
-	apiextensionsv1 "github.com/crossplane/crossplane/v2/apis/apiextensions/v1"
-	"github.com/crossplane/crossplane/v2/test/e2e"
-	"github.com/crossplane/crossplane/v2/test/e2e/config"
-	"github.com/crossplane/crossplane/v2/test/e2e/funcs"
-)
-
-// LabelAreaDiff is applied to all features pertaining to the diff command.
-const LabelAreaDiff = "diff"
-
-// LabelCrossplaneVersion represents the crossplane version compatibility of a test.
-const LabelCrossplaneVersion = "crossplane-version"
-
-// Crossplane version values.
-const (
-	CrossplaneVersionMain       = "main"
-	CrossplaneVersionRelease120 = "release-1.20"
-)
-
-// runCrossplaneDiff runs the crossplane diff command with the specified subcommand on the provided resources.
-// It returns the output and any error encountered.
-func runCrossplaneDiff(t *testing.T, c *envconf.Config, binPath, subcommand string, resourcePaths ...string) (string, string, error) {
-	t.Helper()
-
-	// When E2E_DUMP_EXPECTED is set, run without verbose mode to get clean output for expected files
-	args := []string{subcommand, "--timeout=2m"}
-	if os.Getenv("E2E_DUMP_EXPECTED") == "" {
-		args = append([]string{"--verbose"}, args...)
-	}
-	args = append(args, resourcePaths...)
-
-	t.Logf("Running command: %s %s", binPath, strings.Join(args, " "))
-	cmd := exec.Command(binPath, args...)
-
-	cmd.Env = append(os.Environ(), "KUBECONFIG="+c.KubeconfigFile())
-	t.Logf("ENV: %s KUBECONFIG=%s", binPath, c.KubeconfigFile())
-
-	// Capture standard output and error
-	var stdout, stderr bytes.Buffer
-
-	cmd.Stdout = &stdout
-	cmd.Stderr = &stderr
-
-	// Run the command
-	err := cmd.Run()
-
-	return stdout.String(), stderr.String(), err
-}
-
-// RunXRDiff runs the crossplane xr diff command on the provided resources.
-// It returns the output and any error encountered.
-func RunXRDiff(t *testing.T, c *envconf.Config, binPath string, resourcePaths ...string) (string, string, error) {
-	t.Helper()
-	return runCrossplaneDiff(t, c, binPath, "xr", resourcePaths...)
-}
-
-// RunCompDiff runs the crossplane comp diff command on the provided compositions.
-// It returns the output and any error encountered.
-func RunCompDiff(t *testing.T, c *envconf.Config, binPath string, compositionPaths ...string) (string, string, error) {
-	t.Helper()
-	return runCrossplaneDiff(t, c, binPath, "comp", compositionPaths...)
-}
-
-// TestDiffNewResourceV2Cluster tests the crossplane diff command against net-new resources in v2-cluster variant.
-func TestDiffNewResourceV2Cluster(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-cluster")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffNewResourceV2Cluster").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanDiffNewResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "new-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "new-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.DeleteResourcesWithPropagationPolicy(setupPath, "*.yaml", metav1.DeletePropagationForeground),
-				funcs.ResourcesDeletedWithin(3*time.Minute, setupPath, "*.yaml"),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffExistingResourceV2Cluster tests the crossplane diff command against existing resources in v2-cluster variant.
-func TestDiffExistingResourceV2Cluster(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-cluster")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingResourceV2Cluster").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xr.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xr.yaml"),
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffExistingResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "modified-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteResources", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", clusterNopList)).
-			Feature(),
-	)
-}
-
-// TestDiffNewResourceV2Namespaced tests the crossplane diff command against net-new resources in v2-namespaced variant.
-func TestDiffNewResourceV2Namespaced(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-namespaced")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffNewResourceV2Namespaced").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanDiffNewResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "new-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "new-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.DeleteResourcesWithPropagationPolicy(setupPath, "*.yaml", metav1.DeletePropagationForeground),
-				funcs.ResourcesDeletedWithin(3*time.Minute, setupPath, "*.yaml"),
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffExistingResourceV2Namespaced tests the crossplane diff command against existing resources in v2-namespaced variant.
-func TestDiffExistingResourceV2Namespaced(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-namespaced")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingResourceV2Namespaced").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xr.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xr.yaml"),
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffExistingResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "modified-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteResources", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", nsNopList),
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffNewResourceV1 tests the crossplane diff command against net-new resources in v1 variant.
-func TestDiffNewResourceV1(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v1")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffNewResourceV1").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionRelease120).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanDiffNewResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "new-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "new-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.DeleteResourcesWithPropagationPolicy(setupPath, "*.yaml", metav1.DeletePropagationForeground),
-				funcs.ResourcesDeletedWithin(3*time.Minute, setupPath, "*.yaml"),
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffExistingResourceV1 tests the crossplane diff command against existing resources in v1 variant.
-func TestDiffExistingResourceV1(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v1")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingResourceV1").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionRelease120).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xr.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xr.yaml"),
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffExistingResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "modified-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteResources", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				func(ctx context.Context, t *testing.T, e *envconf.Config) context.Context {
-					t.Helper()
-					// default to `main` variant
-					nopList := clusterNopList
-
-					// we should only ever be running with one version label
-					if slices.Contains(e.Labels()[LabelCrossplaneVersion], CrossplaneVersionRelease120) {
-						nopList = v1NopList
-					}
-
-					funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", nopList)(ctx, t, e)
-
-					return ctx
-				},
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffNewClaim tests the crossplane diff command against net-new claims with v1 XRDs.
-func TestDiffNewClaim(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v1-claim")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffNewClaim").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionRelease120).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanDiffNewClaim", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "new-claim.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "new-claim.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.DeleteResourcesWithPropagationPolicy(setupPath, "*.yaml", metav1.DeletePropagationForeground),
-				funcs.ResourcesDeletedWithin(3*time.Minute, setupPath, "*.yaml"),
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffExistingClaim tests the crossplane diff command against existing claims with v1 XRDs.
-func TestDiffExistingClaim(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v1-claim")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingClaim").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionRelease120).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateClaim", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-claim.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-claim.yaml"),
-				// Claims get their status from the backing XR, so wait for the XR to be available
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-claim.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffExistingClaim", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "modified-claim.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-claim.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteClaim", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-claim.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-claim.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				func(ctx context.Context, t *testing.T, e *envconf.Config) context.Context {
-					t.Helper()
-					// default to `main` variant
-					nopList := clusterNopList
-
-					// we should only ever be running with one version label
-					if slices.Contains(e.Labels()[LabelCrossplaneVersion], CrossplaneVersionRelease120) {
-						nopList = v1NopList
-					}
-
-					funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", nopList)(ctx, t, e)
-
-					return ctx
-				},
-				funcs.ResourceDeletedWithin(3*time.Minute, &k8sapiextensionsv1.CustomResourceDefinition{
-					TypeMeta:   metav1.TypeMeta{Kind: "CustomResourceDefinition", APIVersion: "apiextensions.k8s.io/v1"},
-					ObjectMeta: metav1.ObjectMeta{Name: "nopresources.diff.example.org"},
-				}),
-			)).
-			Feature(),
-	)
-}
-
-var v1NopList = composed.NewList(
-	composed.FromReferenceToList(corev1.ObjectReference{
-		APIVersion: "nop.crossplane.io/v1alpha1",
-		Kind:       "NopResource",
-	}),
-)
-
-var nsNopList = composed.NewList(
-	composed.FromReferenceToList(corev1.ObjectReference{
-		APIVersion: "nop.crossplane.io/v1alpha1",
-		Kind:       "NopResource",
-		Namespace:  "default",
-	}))
-
-var clusterNopList = composed.NewList(composed.FromReferenceToList(corev1.ObjectReference{
-	APIVersion: "nop.crossplane.io/v1alpha1",
-	Kind:       "ClusterNopResource",
-}))
-
-// Regular expressions to match the dynamic parts.
-var (
-	resourceNameRegex              = regexp.MustCompile(`(existing-resource)-[a-z0-9]{5,}(?:-nop-resource)?`)
-	compResourceNameRegex          = regexp.MustCompile(`(test-comp-resource)-[a-z0-9]{5,}`)
-	fanoutResourceNameRegex        = regexp.MustCompile(`(test-fanout-resource-\d{2})-[a-z0-9]{5,}`)
-	getComposedResourceNameRegex   = regexp.MustCompile(`(test-getcomposed-resource)-[a-z0-9]{5,}`)
-	claimNameRegex                 = regexp.MustCompile(`(test-claim)-[a-z0-9]{5,}(?:-[a-z0-9]{5,})?`)
-	claimCompositionRevisionRegex  = regexp.MustCompile(`(xnopclaims\.claim\.diff\.example\.org)-[a-z0-9]{7,}`)
-	compositionRevisionRegex       = regexp.MustCompile(`(xnopresources\.(cluster\.|legacy\.)?diff\.example\.org)-[a-z0-9]{7,}`)
-	nestedCompositionRevisionRegex = regexp.MustCompile(`(child-nop-composition|parent-nop-composition)-[a-z0-9]{7,}`)
-	ansiEscapeRegex                = regexp.MustCompile(`\x1b\[[0-9;]*m`)
-)
-
-// NormalizeLine replaces dynamic parts with fixed placeholders.
-func normalizeLine(line string) string {
-	// Remove ANSI escape sequences
-	line = ansiEscapeRegex.ReplaceAllString(line, "")
-
-	// Replace resource names with random suffixes
-	line = resourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = compResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = fanoutResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = getComposedResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = claimNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-
-	// Replace composition revision refs with random hash
-	line = compositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-	line = claimCompositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-	line = nestedCompositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-
-	// Trim trailing whitespace
-	line = strings.TrimRight(line, " ")
-
-	return line
-}
-
-// normalizeLineKeepAnsi replaces dynamic parts with fixed placeholders while preserving ANSI color codes.
-// Used when writing expected output files that should maintain color codes for human readability.
-func normalizeLineKeepAnsi(line string) string {
-	// Replace resource names with random suffixes
-	line = resourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = compResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = fanoutResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = getComposedResourceNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-	line = claimNameRegex.ReplaceAllString(line, "${1}-XXXXX")
-
-	// Replace composition revision refs with random hash
-	line = compositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-	line = claimCompositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-	line = nestedCompositionRevisionRegex.ReplaceAllString(line, "${1}-XXXXXXX")
-
-	// Trim trailing whitespace
-	line = strings.TrimRight(line, " ")
-
-	return line
-}
-
-// parseStringContent converts a string to raw and normalized lines.
-func parseStringContent(content string) ([]string, []string) {
-	var (
-		rawLines        []string
-		normalizedLines []string
-	)
-
-	scanner := bufio.NewScanner(strings.NewReader(content))
-	for scanner.Scan() {
-		rawLine := scanner.Text()
-		rawLines = append(rawLines, rawLine)
-		normalizedLines = append(normalizedLines, normalizeLine(rawLine))
-	}
-
-	return rawLines, normalizedLines
-}
-
-// AssertDiffMatchesFile compares a diff output with an expected file, ignoring dynamic parts.
-// When E2E_DUMP_EXPECTED environment variable is set, it writes the normalized actual output
-// to the expected file path instead of performing assertions.
-func assertDiffMatchesFile(t *testing.T, actual, expectedSource, log string) {
-	t.Helper()
-
-	actualRaw, actualNormalized := parseStringContent(actual)
-
-	// If E2E_DUMP_EXPECTED is set, write the normalized output to the expected file
-	if os.Getenv("E2E_DUMP_EXPECTED") != "" {
-		// Reconstruct the actual output as a string (with normalization applied, keeping ANSI codes)
-		var normalizedOutput strings.Builder
-		for _, line := range actualRaw {
-			// Normalize each line before writing, but keep ANSI color codes for readability
-			normalizedOutput.WriteString(normalizeLineKeepAnsi(line))
-			normalizedOutput.WriteString("\n")
-		}
-
-		// Create directory if it doesn't exist
-		dir := filepath.Dir(expectedSource)
-		if err := os.MkdirAll(dir, 0o755); err != nil {
-			t.Fatalf("Failed to create directory for expected file: %v", err)
-		}
-
-		// Write the normalized output
-		if err := os.WriteFile(expectedSource, []byte(normalizedOutput.String()), 0o644); err != nil {
-			t.Fatalf("Failed to write expected file: %v", err)
-		}
-
-		t.Logf("Wrote normalized output to %s", expectedSource)
-		return
-	}
-
-	expected, err := os.ReadFile(expectedSource)
-	if err != nil {
-		t.Fatalf("Failed to read expected file: %v", err)
-	}
-
-	expectedRaw, expectedNormalized := parseStringContent(string(expected))
-
-	if len(expectedNormalized) != len(actualNormalized) {
-		t.Errorf("Line count mismatch: expected %d lines in %s, got %d lines in output",
-			len(expectedNormalized), expectedSource,
-			len(actualNormalized))
-	}
-
-	maxLines := max(len(actualNormalized), len(expectedNormalized))
-
-	failed := false
-
-	for i := range maxLines {
-		if i >= len(expectedNormalized) {
-			t.Errorf("Line %d: Extra line in output: %s",
-				i+1, makeStringReadable(actualRaw[i]))
-
-			failed = true
-
-			continue
-		}
-
-		if i >= len(actualNormalized) {
-			t.Errorf("Line %d: Missing line in output: %s",
-				i+1, makeStringReadable(expectedRaw[i]))
-
-			failed = true
-
-			continue
-		}
-
-		if expectedNormalized[i] != actualNormalized[i] {
-			// ignore white space at end of lines
-			// if strings.TrimRight(expectedNormalized[i], " ") == strings.TrimRight(actualNormalized[i], " ") {
-			//	continue
-			//}
-			rawExpected := ""
-			if i < len(expectedRaw) {
-				rawExpected = expectedRaw[i]
-			}
-
-			rawActual := ""
-			if i < len(actualRaw) {
-				rawActual = actualRaw[i]
-			}
-
-			t.Errorf("Line %d mismatch:\n  Expected: %s\n  Actual:   %s\n\n"+
-				"Raw Values (with escape chars visible):\n"+
-				"  Expected Raw: %s\n"+
-				"  Actual Raw:   %s",
-				i+1,
-				expectedNormalized[i],
-				actualNormalized[i],
-				makeStringReadable(rawExpected),
-				makeStringReadable(rawActual))
-
-			failed = true
-		}
-	}
-
-	if failed {
-		t.Errorf("###### Manifest (actual): \n%s\n", actual)
-		t.Errorf("------------------------------------------------------------------")
-		t.Errorf("###### Manifest (expected): \n%s\n", string(expected))
-
-		t.Fatalf("Log output:\n%s", log)
-	}
-}
-
-// makeStringReadable converts a string to a form where all non-printable characters
-// (including ANSI escape codes) are converted to visible escape sequences.
-func makeStringReadable(s string) string {
-	var result strings.Builder
-
-	for _, r := range s {
-		switch {
-		case r == '\x1b':
-			result.WriteString("\\x1b")
-		case r == '\n':
-			result.WriteString("\\n")
-		case r == '\r':
-			result.WriteString("\\r")
-		case r == '\t':
-			result.WriteString("\\t")
-		case r == ' ':
-			result.WriteString("\\space")
-		case !unicode.IsPrint(r):
-			result.WriteString(fmt.Sprintf("\\x%02x", r))
-		default:
-			result.WriteRune(r)
-		}
-	}
-
-	return result.String()
-}
-
-// TestDiffExistingComposition tests the crossplane comp diff command against existing XRs in the cluster.
-func TestDiffExistingComposition(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "comp")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingComposition").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xr.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xr.yaml"),
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffComposition", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunCompDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "updated-composition.yaml"))
-				if err != nil {
-					t.Fatalf("Error running comp diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteExistingXR", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", clusterNopList)).
-			Feature(),
-	)
-}
-
-// TestCompDiffLargeFanout tests the comp diff command with a large number of XRs (15) using the same composition.
-// This validates that:
-// 1. Container reuse works correctly across many XRs
-// 2. Cleanup happens properly after processing all XRs
-// 3. Diff output is generated for all XRs
-// 4. No performance issues or failures with large fanout.
-func TestCompDiffLargeFanout(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "comp-fanout")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("CompDiffLargeFanout").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeLarge).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXRs", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xrs.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xrs.yaml"),
-				funcs.ResourcesHaveConditionWithin(3*time.Minute, manifests, "existing-xrs.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffCompositionWithLargeFanout", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunCompDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "updated-composition.yaml"))
-				if err != nil {
-					t.Fatalf("Error running comp diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xrs.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteExistingXRs", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xrs.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xrs.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", clusterNopList)).
-			Feature(),
-	)
-}
-
-// TestDiffCompositionWithGetComposedResource tests the comp diff command with a composition
-// that uses getComposedResource template function to reference observed resources.
-// This test verifies that ObservedResources are correctly populated during rendering.
-func TestDiffCompositionWithGetComposedResource(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "comp-getcomposed")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffCompositionWithGetComposedResource").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-xr.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, manifests, "existing-xr.yaml"),
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffCompositionWithGetComposedResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunCompDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "updated-composition.yaml"))
-				if err != nil {
-					t.Fatalf("Error running comp diff command: %v\nLog output:\n%s", err, log)
-				}
-
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteExistingXR", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", clusterNopList)).
-			Feature(),
-	)
-}
-
-// TestDiffConcurrentDirectory tests issue #59 - concurrent function startup failures
-// when processing multiple XR files from a directory with a composition using multiple functions.
-func TestDiffConcurrentDirectory(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-concurrent-dir")
-	setupPath := filepath.Join(manifests, "setup")
-	xrsPath := filepath.Join(manifests, "xrs")
-
-	environment.Test(t,
-		features.New("TestDiffConcurrentDirectory").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeLarge).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanProcessDirectoryWithMultipleXRs", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				// Get all XR files from the directory
-				xrFiles, err := filepath.Glob(filepath.Join(xrsPath, "*.yaml"))
-				if err != nil {
-					t.Fatalf("Failed to find XR files: %v", err)
-				}
-
-				if len(xrFiles) != 21 {
-					t.Fatalf("Expected 21 XR files, found %d", len(xrFiles))
-				}
-
-				// Run diff on all XR files - this tests concurrent function processing
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", xrFiles...)
-
-				// Always log output for debugging
-				t.Logf("crossplane-diff stdout: %s", output)
-				t.Logf("crossplane-diff stderr: %s", log)
-
-				if err != nil {
-					t.Fatalf("Error running diff command: %v", err)
-				}
-
-				// Verify we processed all XRs - each XR creates 1 NopResource
-				// With 21 XRs, we should see 21 "+++ NopResource/" lines in the output
-				addedCount := strings.Count(output, "+++ NopResource/")
-				if addedCount != 21 {
-					t.Errorf("Expected 21 NopResource additions, found %d", addedCount)
-				}
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.DeleteResourcesWithPropagationPolicy(setupPath, "*.yaml", metav1.DeletePropagationForeground),
-				funcs.ResourcesDeletedWithin(3*time.Minute, setupPath, "*.yaml"),
-			)).
-			Feature(),
-	)
-}
-
-// end of file
-
-// TestDiffNewNestedResourceV2 tests the crossplane diff command against net-new nested XR resources in v2 variant.
-func TestDiffNewNestedResourceV2(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-nested")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffNewNestedResourceV2").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "parent-definition.yaml", apiextensionsv1.WatchingComposite()),
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "child-definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			Assess("CanDiffNewNestedResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "new-parent-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "new-parent-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", nsNopList),
-			)).
-			Feature(),
-	)
-}
-
-// TestDiffExistingNestedResourceV2 tests the crossplane diff command against existing nested XR resources in v2 variant.
-func TestDiffExistingNestedResourceV2(t *testing.T) {
-	imageTag := strings.Split(environment.GetCrossplaneImage(), ":")[1]
-	manifests := filepath.Join("test/e2e/manifests/beta/diff", imageTag, "v2-nested")
-	setupPath := filepath.Join(manifests, "setup")
-	expectPath := filepath.Join(manifests, "expect")
-
-	environment.Test(t,
-		features.New("DiffExistingNestedResourceV2").
-			WithLabel(e2e.LabelArea, LabelAreaDiff).
-			WithLabel(e2e.LabelSize, e2e.LabelSizeSmall).
-			WithLabel(config.LabelTestSuite, config.TestSuiteDefault).
-			WithLabel(LabelCrossplaneVersion, CrossplaneVersionMain).
-			WithSetup("CreatePrerequisites", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, setupPath, "*.yaml"),
-				funcs.ResourcesCreatedWithin(30*time.Second, setupPath, "*.yaml"),
-			)).
-			WithSetup("PrerequisitesAreReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "parent-definition.yaml", apiextensionsv1.WatchingComposite()),
-				funcs.ResourcesHaveConditionWithin(1*time.Minute, setupPath, "child-definition.yaml", apiextensionsv1.WatchingComposite()),
-			)).
-			WithSetup("CreateExistingXR", funcs.AllOf(
-				funcs.ApplyResources(e2e.FieldManager, manifests, "existing-parent-xr.yaml"),
-				funcs.ResourcesCreatedWithin(1*time.Minute, manifests, "existing-parent-xr.yaml"),
-			)).
-			WithSetup("ExistingXRIsReady", funcs.AllOf(
-				funcs.ResourcesHaveConditionWithin(2*time.Minute, manifests, "existing-parent-xr.yaml", xpv1.Available()),
-			)).
-			Assess("CanDiffExistingNestedResource", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
-				t.Helper()
-
-				output, log, err := RunXRDiff(t, c, "./crossplane-diff", filepath.Join(manifests, "modified-parent-xr.yaml"))
-				if err != nil {
-					t.Fatalf("Error running diff command: %v\nLog output:\n%s", err, log)
-				}
-
-				assertDiffMatchesFile(t, output, filepath.Join(expectPath, "existing-parent-xr.ansi"), log)
-
-				return ctx
-			}).
-			WithTeardown("DeleteResources", funcs.AllOf(
-				funcs.DeleteResources(manifests, "existing-parent-xr.yaml"),
-				funcs.ResourcesDeletedWithin(2*time.Minute, manifests, "existing-parent-xr.yaml"),
-			)).
-			WithTeardown("DeletePrerequisites", funcs.AllOf(
-				funcs.ResourcesDeletedAfterListedAreGone(3*time.Minute, setupPath, "*.yaml", nsNopList),
-			)).
-			Feature(),
-	)
-}
